Some small bugs but most of the expander works
