--- out1
@@ -1,1 +1,0 @@
-heya this is my testing file
